+++ ChangeLog.md
@@ -1,2 +1,3 @@
 ##0.0.1(未发布)
 1. Update: remove default_bookmarks content and set a wrong UA
+2. Update: remove UA, Search and HomePage pref